### 2023-05-31 #5508 @editorialbot
[PRE REVIEW]: pvOps: a Python package for empirical analysis of photovoltaic field data
**Submitting author:** @klbonne (<a href="http://orcid.org/0009-0006-2383-1634">Kirk Bonney</a>)
**Repository:** https://github.com/sandialabs/pvOps
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.1.9
