### 2023-06-20 02adf98
add helper functions and reference dict
