### 2023-06-21 3f1f128
clean up text, tokenize, add new labels from regex function
