+++ pvops/text/classify_regex.py
@@ -1,4 +1,5 @@
 import pandas as pd
+from nltk.tokenize import word_tokenize
 
 # TODO: compare input, output of classify module
 
@@ -7,8 +8,9 @@ import pandas as pd
 2. run through regex function
 3. get expected output
 """
-LABEL_COLUMN = 'Problem Level'
-NOTES_COLUMN = 'Problem Description'
+LABEL_COLUMN = 'equipment_label'
+REGEX_LABEL_COLUMN = 'regex_equipment_label'
+NOTES_COLUMN = 'notes'
 
 # first pass of equipment dictionary is based on
 # https://www.osti.gov/servlets/purl/1872704
@@ -29,10 +31,13 @@ EQUIPMENT_DICT = {'combiner': ['combiner', 'comb', 'cb'],
                   }
 
 def get_sample_data():
-    filename = '~/data/charity/doe_data/sm_logs_notes_only.csv'
+    filename = '~/data/charity/doe_data/sm_logs_notes_cleaned.csv'
     cols = [NOTES_COLUMN, LABEL_COLUMN] # TODO: add more columns of interest later
     df = pd.read_csv(filename)
+    # drop any logs without notes or resolution notes
+    df = df[~df[NOTES_COLUMN].isna()]
     return df[cols]
+
 
 def get_all_keywords_of_interest(list_of_txt, custom_dict=None):
     """
@@ -61,12 +66,15 @@ def get_all_keywords_of_interest(list_of_txt, custom_dict=None):
     return included_equipment
 
 class Example:
-    def __init__(self, df, LABEL_COLUMN):
+    def __init__(self, df, REGEX_LABEL_COLUMN):
         """
         where df is expected input
         """
-        self.LABEL_COLUMN = LABEL_COLUMN
+        self.REGEX_LABEL_COLUMN = REGEX_LABEL_COLUMN
         self.df = df
+
+        # tokenize notes to words
+        self.df[NOTES_COLUMN] = self.df[NOTES_COLUMN].apply(word_tokenize)
 
     def add_equipment_labels(self):
         """
@@ -74,13 +82,14 @@ class Example:
         dataframe with additional 'equipment_label' column
         dependent on entries in EQUIPMENT_DICT
         """
-        self.df[LABEL_COLUMN] = self.df[NOTES_COLUMN].apply(get_all_keywords_of_interest)
+        self.df[REGEX_LABEL_COLUMN] = self.df[NOTES_COLUMN].apply(get_all_keywords_of_interest)
 
         # each multi-category now in its own row
         # some logs have multiple equipment issues
-        self.df = self.df.explode(LABEL_COLUMN)
+        self.df = self.df.explode(REGEX_LABEL_COLUMN)
 
 if __name__ == "__main__":
     # python -m pvops.text.classify_regex
     df = get_sample_data()
-    e = Example(df=df, LABEL_COLUMN=LABEL_COLUMN)
+    e = Example(df=df, REGEX_LABEL_COLUMN=REGEX_LABEL_COLUMN)
+    e.add_equipment_labels()
