### 2023-06-21 80975b6
move example into its own file, seperate out functions to use
