### 2023-06-21 257d657
clean functions and make generic names
