+++ pvops/text/classify_regex.py
@@ -1,13 +1,9 @@
 
-# TODO: compare input, output of classify module
-
 """
-1. get expected input (CCRE data)
-2. run through regex function
-3. get expected output
+Temporary placeholder to hold classify by search functions
 """
 
-# first pass of equipment dictionary is based on
+# Default equipment dictionary is based on
 # https://www.osti.gov/servlets/purl/1872704
 # TODO: only currently takes single words, not phrases: 'string inverter', 'met station', etc.
 EQUIPMENT_DICT = {'combiner': ['combiner', 'comb', 'cb'],
@@ -25,28 +21,37 @@ EQUIPMENT_DICT = {'combiner': ['combiner', 'comb', 'cb'],
                   'wiring': ['wiring', 'wire', 'wires']
                   }
 
-# TODO: allow .csv input also
+def get_keywords_of_interest(list_of_txt, reference_dict=None):
+    """Find keywords of interest in list of strings from reference dict.
 
-def get_all_keywords_of_interest(list_of_txt, reference_dict=None):
-    """
-    if keywords of interest are in the list of text, return the keyword category
-    for example, if 'inverter' are in the list of text, return ['inverter']
+    If keywords of interest given in a reference dict are in the list of
+    strings, return the keyword category, or categories. For example,
+    if the string 'inverter' is in the list of text, return ['inverter'].
 
     Parameters
     ----------
-    list_of_txt: list
-        list of strings
-    reference_dict: dict
-    
+    list_of_txt : list of str
+        Tokenized text, functionally a list of string values.
+    reference_dict : dict with {'keyword': [list of synonyms]} or None
+        Reference dictionary to search for keywords of interest,
+        in the expected format
+        {'keyword_a':
+            ['keyword_a', 'keyword_a_synonym_0', 'keyword_a_synonym_1', keyword_a_synonym_2', ...],
+         'keyword_b':
+            ['keyword_b', 'keyword_b_synonym_0', 'keyword_b_synonym_1', keyword_b_synonym_2', ...],
+         ...}
+        If None, use default reference dictionary.
+        Note: This function can currently only handle single words, no n-gram functionality.
+
     Returns
     -------
-    included_equipment: list
-        list of strings (included equipment named)
-
+    included_equipment: list of str
+        List of keywords from reference_dict found in list_of_txt, can be more than one value.
     """
     text_to_search = set(list_of_txt)
     
     if reference_dict is None:
+        # use default equipment lookup if no reference dict is given
         reference_dict = EQUIPMENT_DICT
 
     equipment_keywords = set(reference_dict.keys())
@@ -54,15 +59,42 @@ def get_all_keywords_of_interest(list_of_txt, reference_dict=None):
 
     return included_equipment
 
-def add_equipment_labels(self):
-    """
-    manually add labels to mor text logs
-    dataframe with additional 'equipment_label' column
-    dependent on entries in EQUIPMENT_DICT
-    """
-    self.df[REGEX_LABEL_COLUMN] = self.df[NOTES_COLUMN].apply(get_all_keywords_of_interest)
+def add_keyword_labels(df, text_col, new_col, reference_dict=None):
+    """Find keywords of interest in specified column of dataframe, return as new column value.
 
-    # each multi-category now in its own row
-    # some logs have multiple equipment issues
-    self.df = self.df.explode(REGEX_LABEL_COLUMN)
+    If keywords of interest given in a reference dict are in the specified column of the dataframe,
+    return the keyword category, or categories. For example, if the string 'inverter'
+    is in the list of text, return ['inverter'].
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        Dataframe to search for keywords of interest, must include text_col.
+    text_col : str
[…]
