+++ examples/text_classify_regex_example.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from nltk.tokenize import word_tokenize
 
-from pvops.text.classify_regex import add_equipment_labels
+from pvops.text.classify_regex import add_equipment_labels, EQUIPMENT_DICT
 
 LABEL_COLUMN = 'equipment_label'
 REGEX_LABEL_COLUMN = 'regex_equipment_label'
@@ -26,8 +26,15 @@ class Example:
         # tokenize notes to words
         self.df[NOTES_COLUMN] = self.df[NOTES_COLUMN].apply(word_tokenize)
 
+    def add_equipment_labels(self):
+        text_col = NOTES_COLUMN
+        new_col = REGEX_LABEL_COLUMN
+        reference_dict = EQUIPMENT_DICT
+        self.df = add_equipment_labels(self.df, text_col, new_col, reference_dict)
+
 if __name__ == "__main__":
-    # python -m pvops.text.classify_regex
+    # python -m examples.classify_regex_example
     df = get_sample_data()
     e = Example(df=df, REGEX_LABEL_COLUMN=REGEX_LABEL_COLUMN)
     e.add_equipment_labels()
+    import pdb; pdb.set_trace()
