### 2023-06-21 62a32c4
clean up example
