+++ examples/text_classify_regex_example.py
@@ -1,40 +1,51 @@
 import pandas as pd
 from nltk.tokenize import word_tokenize
 
-from pvops.text.classify_regex import add_equipment_labels, EQUIPMENT_DICT
+from pvops.text.classify_regex import add_keyword_labels, EQUIPMENT_DICT
 
 LABEL_COLUMN = 'equipment_label'
-REGEX_LABEL_COLUMN = 'regex_equipment_label'
+NEW_LABEL_COLUMN = 'new_equipment_label'
 NOTES_COLUMN = 'notes'
 
-def get_sample_data():
-    filename = '~/data/charity/doe_data/sm_logs_notes_cleaned.csv'
-    cols = [NOTES_COLUMN, LABEL_COLUMN] # TODO: add more columns of interest later
+def get_sample_data(filename):
+    """Function to read .csv file of data.
+
+    Parameters
+    ----------
+    filename : str
+
+    Returns
+    -------
+    df: pd.DataFrame
+    """
+
     df = pd.read_csv(filename)
-    # drop any logs without notes or resolution notes
-    df = df[~df[NOTES_COLUMN].isna()]
-    return df[cols]
+    df = df[~df[NOTES_COLUMN].isna()] # drop any logs without notes
+    return df[[NOTES_COLUMN, LABEL_COLUMN]]
 
 class Example:
-    def __init__(self, df, REGEX_LABEL_COLUMN):
+    def __init__(self, df):
         """
-        where df is expected input
+        Parameters
+        ----------
+        df : pd.DataFrame
+            Must have columns LABEL_COLUMN, and NOTES_COLUMN
         """
-        self.REGEX_LABEL_COLUMN = REGEX_LABEL_COLUMN
         self.df = df
 
         # tokenize notes to words
         self.df[NOTES_COLUMN] = self.df[NOTES_COLUMN].apply(word_tokenize)
 
     def add_equipment_labels(self):
-        text_col = NOTES_COLUMN
-        new_col = REGEX_LABEL_COLUMN
-        reference_dict = EQUIPMENT_DICT
-        self.df = add_equipment_labels(self.df, text_col, new_col, reference_dict)
+        """Add new equipment labels.
+        """
+        self.df = add_keyword_labels(self.df,
+                                     text_col=NOTES_COLUMN,
+                                     new_col=NEW_LABEL_COLUMN,
+                                     reference_dict=EQUIPMENT_DICT)
 
 if __name__ == "__main__":
-    # python -m examples.classify_regex_example
-    df = get_sample_data()
-    e = Example(df=df, REGEX_LABEL_COLUMN=REGEX_LABEL_COLUMN)
+    # python -m examples.text_classify_regex_example
+    df = get_sample_data(filename='~/data/charity/doe_data/sm_logs_notes_cleaned.csv')
+    e = Example(df)
     e.add_equipment_labels()
-    import pdb; pdb.set_trace()
