### 2023-06-21 a3e7e0b
move reference dict to example_data
