+++ examples/text_classify_regex_example.py
@@ -1,7 +1,9 @@
 import pandas as pd
 from nltk.tokenize import word_tokenize
 
-from pvops.text.classify_regex import add_keyword_labels, EQUIPMENT_DICT
+from pvops.text import utils
+from pvops.text.classify_regex import add_keyword_labels
+from examples.example_data.reference_dict import EQUIPMENT_DICT
 
 LABEL_COLUMN = 'equipment_label'
 NEW_LABEL_COLUMN = 'new_equipment_label'
@@ -21,6 +23,21 @@ def get_sample_data(filename):
 
     df = pd.read_csv(filename)
     df = df[~df[NOTES_COLUMN].isna()] # drop any logs without notes
+
+    # remap assets
+    folder = 'example_data//'
+    asset_remap_filename = 'remappings_asset.csv'
+    REMAPPING_COL_FROM = 'in'
+    REMAPPING_COL_TO = 'out_'
+    remapping_df = pd.read_csv('~/pvOps/examples/example_data/remappings_asset.csv')
+    remapping_col_dict = {
+        'attribute_col': LABEL_COLUMN,
+        'remapping_col_from': REMAPPING_COL_FROM,
+        'remapping_col_to': REMAPPING_COL_TO
+    }
+
+    df = utils.remap_attributes(df, remapping_df, remapping_col_dict, allow_missing_mappings=True)
+
     return df[[NOTES_COLUMN, LABEL_COLUMN]]
 
 class Example:
@@ -49,3 +66,4 @@ if __name__ == "__main__":
     df = get_sample_data(filename='~/data/charity/doe_data/sm_logs_notes_cleaned.csv')
     e = Example(df)
     e.add_equipment_labels()
+    import pdb; pdb.set_trace()
