### 2023-06-21 1376a26
move dict out of functions file
