### 2023-06-23 42a2e44
move add label fucntion to classify
