+++ examples/text_classify_regex_example.py
@@ -1,9 +1,10 @@
 from nltk.tokenize import word_tokenize
 import pandas as pd
-from sklearn.metrics import accuracy_score
+from sklearn.metrics import accuracy_score, ConfusionMatrixDisplay
+import matplotlib.pyplot as plt
 
 from pvops.text import utils
-from pvops.text.classify_regex import add_keyword_labels
+from pvops.text.classify import add_keyword_labels
 from examples.example_data.reference_dict import EQUIPMENT_DICT
 
 LABEL_COLUMN = 'equipment_label'
@@ -74,8 +75,17 @@ class Example:
         msg = f'{label_count:.2%} of entries had a keyword of interest, with {acc_score:.2%} accuracy.'
         print(msg)
 
+        # TODO: clean up visualization function here
+        labels = list(set(self.df[LABEL_COLUMN]).union(set(self.df[NEW_LABEL_COLUMN])))
+        cm_display = ConfusionMatrixDisplay.from_predictions(y_true=self.df[LABEL_COLUMN], y_pred=self.df[NEW_LABEL_COLUMN], display_labels=labels, xticks_rotation='vertical', normalize='true', values_format='.0%')
+        fig, ax = plt.subplots(figsize=(10,10))
+        cm_display.plot(ax=ax)
+
+        cm_display.figure_.savefig('examples/conf_mat.png', dpi=600)
+
 if __name__ == "__main__":
     # python -m examples.text_classify_regex_example
+    # TODO: update this with existing pvops data
     df = get_sample_data(filename='~/data/charity/doe_data/sm_logs_notes_cleaned.csv')
     e = Example(df)
     e.add_equipment_labels()
