### 2023-06-23 7e7aae9
clean up dict
