### 2023-06-23 cb0243e
Updating python setup action in testing workflow
