### 2023-06-26 7b59fba
clean up and use sample data in examples
