+++ examples/text_classify_regex_example.py
@@ -4,6 +4,7 @@ from sklearn.metrics import accuracy_score
 
 from pvops.text import preprocess
 from pvops.text.classify import get_labels_from_keywords
+from pvops.text.visualize import visualize_classification_confusion_matrix
 from examples.example_data.reference_dict import EQUIPMENT_DICT
 
 
@@ -37,17 +38,21 @@ class Example:
         self.om_df = get_labels_from_keywords(self.om_df,
                                               col_dict=self.col_dict,
                                               reference_dict=EQUIPMENT_DICT)
+        
+        # replace 'Other' values with 'Unknown'
+        self.om_df[self.LABEL_COL] = self.om_df[self.LABEL_COL].replace('other', 'unknown')
+        # replace NaN values to use accuracy score
+        self.om_df[[self.LABEL_COL, self.NEW_LABEL_COL]] = self.om_df[[self.LABEL_COL, self.NEW_LABEL_COL]].fillna('unknown')
+
+    def plot_confusion_matrix(self):
+        output_filepath = 'examples/example_conf_mat.png'
+        visualize_classification_confusion_matrix(self.om_df, self.col_dict, output_filepath)
 
     def get_metrics(self):
         """Get accuracy measures and count metrics.
         """
         # entries with some keyword over interest, over all entries
         label_count = self.om_df[self.NEW_LABEL_COL].count() / len(self.om_df)
-
-        # replace 'Other' values with 'Unknown'
-        self.om_df[self.LABEL_COL] = self.om_df[self.LABEL_COL].replace('other', 'unknown')
-        # replace NaN values to use accuracy score
-        self.om_df[[self.LABEL_COL, self.NEW_LABEL_COL]] = self.om_df[[self.LABEL_COL, self.NEW_LABEL_COL]].fillna('unknown')
 
         acc_score = accuracy_score(y_true=self.om_df[self.NEW_LABEL_COL], y_pred=self.om_df[self.NEW_LABEL_COL])
 
@@ -71,4 +76,5 @@ if __name__ == "__main__":
 
     e = Example(om_df=om_df, col_dict=col_dict)
     e.add_equipment_labels()
+    e.plot_confusion_matrix()
     e.get_metrics()
