### 2023-06-26 f008346
add more detail to dox string
