### 2023-06-26 73ba6a6
clean up function to expect predicted_col
