### 2023-06-26 443d6ee
clean up example
