### 2023-06-26 63a74ae
clean up sample data
