+++ examples/example_data/reference_dict.py
@@ -15,3 +15,14 @@ EQUIPMENT_DICT = {'combiner': ['combiner', 'comb', 'cb'],
                   'transformer': ['transformer', 'xfmr'],
                   'wiring': ['wiring', 'wire', 'wires']
                   }
+
+PV_TERMS_DICT = {'communication': ['comm',],
+                 'energy': ['energy', 'kwh', 'mwh',],
+                 'grid': ['grid', 'curtailment', 'curtail', 'poi',],
+                 'outage': ['outage', 'offline',],
+                 'solar': ['solar', 'pv', 'photovoltaic',],
+                 'system': ['system', 'site', 'farm', 'project',],
+                 'make_model': ['sma',], # TODO: use the equipment database for this
+                 'corrective_maintence': ['cm',],
+                 'preventative_maintence': ['pm',]
+                 }
