### 2023-06-27 0990c61
add text replacement step
