+++ examples/text_classify_regex_example.py
@@ -93,7 +93,7 @@ if __name__ == "__main__":
 
     # preprocessing steps
     om_df[col_dict['attribute_col']] = om_df.apply(lambda row: row[col_dict['attribute_col']].lower(), axis=1)
-    equipment_dict = {k:EQUIPMENT_DICT[k] for k in EQUIPMENT_DICT if len(EQUIPMENT_DICT[k]) > 1}
+    equipment_dict = {k: EQUIPMENT_DICT[k] for k in EQUIPMENT_DICT if len(EQUIPMENT_DICT[k]) > 1}
     pv_reference_dict = {**equipment_dict, **PV_TERMS_DICT}
     om_df = preprocess.preprocessor(om_df, lst_stopwords=[], col_dict=col_dict, pv_reference_dict=pv_reference_dict, print_info=False, extract_dates_only=False)
 
