### 2023-07-06 57cc1d1
remove example .py file in favor of jupyter notebook
