+++ pvops/text/preprocess.py
@@ -6,9 +6,6 @@ import datefinder
 import traceback
 from datetime import datetime, timedelta
 
-from examples.example_data.reference_dict import EQUIPMENT_DICT, PV_TERMS_DICT
-from pvops.text.utils import replace_word_with_pv_term
-
 
 try:
     nltk.data.find('tokenizers/punkt')
@@ -16,7 +13,7 @@ except LookupError:
     nltk.download('punkt')
 
 def preprocessor(
-    om_df, lst_stopwords, col_dict, pv_reference_dict=None, print_info=False, extract_dates_only=False
+    om_df, lst_stopwords, col_dict, print_info=False, extract_dates_only=False
 ):
     """Preprocessing function which processes the raw text data into processed text data and extracts dates
 
@@ -95,9 +92,8 @@ def preprocessor(
 
         if not extract_dates_only:
             try:
-                out_stop = text_remove_numbers_stopwords(document, lst_stopwords)
-                out_replace = text_replace_key_terms(out_stop, pv_reference_dict)
-                clean_corpus.append(out_replace)
+                out_ = text_remove_numbers_stopwords(document, lst_stopwords)
+                clean_corpus.append(out_)
             except:
                 print(traceback.format_exc())
                 clean_corpus.append("")
@@ -455,34 +451,8 @@ def text_remove_numbers_stopwords(document, lst_stopwords):
 
     return document
 
-def text_replace_key_terms(document, pv_reference_dict=None):
-    """Conduct final processing steps after date extraction
 
-    Parameters
-    ----------
-    document : str
-        String representation of a document
-    pv_reference_dict : dict or None
-        {most_common_term: list of other terms for the most_common_term}
-
-    Returns
-    -------
-    string
-        string of processed document
-    """
-    if pv_reference_dict is None:
-        # for equipment, only keep dict entries where there are more than one way to reference a term
-        equipment_dict = {k: EQUIPMENT_DICT[k] for k in EQUIPMENT_DICT if len(EQUIPMENT_DICT[k]) > 1}
-        pv_reference_dict = {**equipment_dict, **PV_TERMS_DICT}
-
-    # remove all spaces
-    document_tok = nltk.word_tokenize(document)
-    document = [replace_word_with_pv_term(word, pv_reference_dict) for word in document_tok]
-    document = " ".join(document)
-
-    return document
-
-def get_keywords_of_interest(document_tok, reference_dict=EQUIPMENT_DICT):
+def get_keywords_of_interest(document_tok, reference_df):
     """Find keywords of interest in list of strings from reference dict.
 
     If keywords of interest given in a reference dict are in the list of
@@ -509,9 +479,13 @@ def get_keywords_of_interest(document_tok, reference_dict=EQUIPMENT_DICT):
     included_equipment: list of str
         List of keywords from reference_dict found in list_of_txt, can be more than one value.
     """
-    text_to_search = set(document_tok)
-
-    equipment_keywords = set(reference_dict.keys())
-    included_equipment = list(text_to_search.intersection(equipment_keywords))
-
-    return included_equipment
+    REMAPPING_COL_FROM = 'in'
+    REMAPPING_COL_TO = 'out_'
+    reference_dict = dict(
+        zip(reference_df[REMAPPING_COL_FROM], reference_df[REMAPPING_COL_TO])
+    )
+    
+    # keywords of interest
+    overlap_keywords = reference_dict.keys() & document_tok
+    included_keywords = list({reference_dict[x] for x in overlap_keywords})
+    return included_keywords
