### 2023-07-06 e8d95a3
update docstrings
