+++ pvops/text/classify.py
@@ -190,12 +190,12 @@ def classification_deployer(
 
     return pd.concat(rows, axis=1).T, best_gs_instance.best_estimator_
 
-def get_attributes_from_keywords(om_df, col_dict, reference_df):
+def get_attributes_from_keywords(om_df, col_dict, reference_df, reference_col_dict):
     """Find keywords of interest in specified column of dataframe, return as new column value.
 
-    If keywords of interest given in a reference dict are in the specified column of the dataframe,
-    return the keyword category, or categories. For example, if the string 'inverter'
-    is in the list of text, return ['inverter'].
+    If keywords of interest given in a reference dataframe are in the specified column of the
+    dataframe, return the keyword category, or categories.
+    For example, if the string 'inverter' is in the list of text, return ['inverter'].
 
     Parameters
     ----------
@@ -206,16 +206,20 @@ def get_attributes_from_keywords(om_df, col_dict, reference_df):
 
         - data : string, should be assigned to associated column which stores the tokenized text logs
         - predicted_col : string, will be used to create keyword search label column
-    reference_dict : dict of {'keyword': [list of synonyms]} or None
-        Reference dictionary to search for keywords of interest,
-        in the expected format
-        {'keyword_a':
-            ['keyword_a', 'keyword_a_synonym_0', 'keyword_a_synonym_1', keyword_a_synonym_2', ...],
-         'keyword_b':
-            ['keyword_b', 'keyword_b_synonym_0', 'keyword_b_synonym_1', keyword_b_synonym_2', ...],
-         ...}
-        If None, use default reference dictionary.
+    reference_df : DataFrame
+        Holds columns that define the reference dictionary to search for keywords of interest,
         Note: This function can currently only handle single words, no n-gram functionality.
+    reference_col_dict : dict of {str : str}
+        A dictionary that contains the column names that describes how
+        referencing is going to be done
+
+        - reference_col_from : string, should be assigned to
+          associated column name in reference_df that are possible input reference values
+          Example: pd.Series(['inverter', 'invert', 'inv'])
+        - reference_col_to : string, should be assigned to
+          associated column name in reference_df that are the output reference values
+          of interest
+          Example: pd.Series(['inverter', 'inverter', 'inverter'])
 
     Returns
     -------
@@ -223,7 +227,9 @@ def get_attributes_from_keywords(om_df, col_dict, reference_df):
         Input df with new_col added, where each found keyword is its own row, may result in
         duplicate rows if more than one keywords of interest was found in text_col.
     """
-    om_df[col_dict['predicted_col']] = om_df[col_dict['data']].apply(get_keywords_of_interest, reference_df=reference_df)
+    om_df[col_dict['predicted_col']] = om_df[col_dict['data']].apply(get_keywords_of_interest,
+                                                                     reference_df=reference_df,
+                                                                     reference_col_dict=reference_col_dict)
 
     # each multi-category now in its own row, some logs have multiple equipment issues
     multiple_keywords_df = om_df[om_df[col_dict['predicted_col']].str.len() > 1]
