+++ pvops/text/utils.py
@@ -1,6 +1,7 @@
 import pandas as pd
 import numpy as np
 
+
 def remap_attributes(om_df, remapping_df, remapping_col_dict,
                      allow_missing_mappings=False, print_info=False):
     """A utility function which remaps the attributes of om_df using columns
