+++ pvops/text/classify.py
@@ -10,7 +10,6 @@ import copy
 
 from pvops.text.preprocess import get_keywords_of_interest
 
-
 def classification_deployer(
     X,
     y,
