### 2023-07-07 a4de9a4
Keyword search in text from reference dictionary (#46)
* init file and pull sample data

* add helper functions and reference dict

* clean up text, tokenize, add new labels from regex function

* move example into its own file, seperate out functions to use

* clean functions and make generic names

* clean up example

* move reference dict to example_data

* remove pdb

* move dict out of functions file

* missed import

* add metrics to measure accuracy

* rename other category to unknown for fair comparison

* move keyword search to preprocess step, util

* move add label fucntion to class
[…]
