### 2023-08-10 #5755 @editorialbot
[REVIEW]: pvOps: a Python package for empirical analysis of photovoltaic field data
**Submitting author:** @kbonney (<a href="http://orcid.org/0009-0006-2383-1634">Kirk Bonney</a>)
**Repository:** https://github.com/sandialabs/pvOps
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.3.0
