### 2023-08-14 sandialabs/pvOps#52 @FlorianK13
[Joss Review] Extend caption of Figure 1
Please extend the caption of figure 1.

This comment is part of a JOSS Review: https://github.com/openjournals/joss-reviews/issues/5755
