### 2023-08-28 #5755 @kbonney
> Hi @AdamRJensen, just a reminder to get your review started when you can!
>
> And @kbonney (I just noticed your GH handle was wrong in the top comment here; I just fixed that), any updates on progress in response to the remaining issues raised by @FlorianK13?

We'll start clearing up those issues @rkurchin.
I also have a question for you regarding the formatting of the JOSS document. We had difficulties getting our html-style table to format correctly in the document. Our temporary solution was to use a screen-shot of the properly formatted table, but ideally this would be generated by the JOSS document creation procedure from source. Who can I reach out to for troubleshooting support for this sort of issue?
