### 2023-08-28 sandialabs/pvOps#52 @kbonney
Figure 1 isn't really intended to be a figure. We want to use HTML to render this into the document directly. Pasting in as a screenshot was a temporary fix due to issues with the JOSS document building process. See https://github.com/openjournals/joss-reviews/issues/5755#issuecomment-1696313851.
If the rendered table still comes with a caption, we will make sure it is more descriptive than "table".
