### 2023-08-30 #5755 @kbonney
Yes @arfon. Our table includes bulleted lists inside the entries of the last two columns. To get this formatting right in markdown we were using \<br> tags to force linebreaks, otherwise the bullet points run together. Using these tags renders how we want in standard MD+HTML, but when we run the code through the joss draft workflow the \<br> tags seem to be ignored. You can see this happen in the document generated here: https://github.com/kbonney/pvOps/actions/runs/6026505579.
