### 2023-09-12 sandialabs/pvOps#55 @AdamRJensen
[JOSS review] Consider changing copyright/author
I noticed that the copyright/author on the readthedocs website is different from the main author of the JOSS paper. Perhaps it should be considered changing this to something more generic, e.g., "pvOps Developers" or to Sandia who is stated in the copyright file?

Note, I have no preference for what is provided, I merely wanted to bring attention to this to ensure it is still correct.

This is part of a JOSS review: https://github.com/openjournals/joss-reviews/issues/5755

https://github.com/sandialabs/pvOps/blob/fe6e6579239ab161908fc5b0d3819b720c161f3c/docs/conf.py#L21-L22

![image](https://github.com/sandialabs/pvOps/assets/39184289/0ad1e434-83b8-4a0c-bdf6-523b23d77f71)
