### 2023-09-12 sandialabs/pvOps#56 @AdamRJensen
[JOSS review] Make .__version__ available
It's quasi-standard for Python packages to be able to return the package version using ``.__version__``

This doesn't seem possible using the pvops package at the moment:

    pvops.__version__

    Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    AttributeError: module 'pvops' has no attribute '__version__'

Having users easily retrieve the package version is important in science, thus I think this is somewhat important to implement.

This issue is part of a JOSS review: https://github.com/openjournals/joss-reviews/issues/5755
