### 2023-09-12 sandialabs/pvOps#59 @AdamRJensen
[JOSS review] Avoid separate doc requirements.txt
I'm confused as to why there is a separate [requirements.txt](https://github.com/sandialabs/pvOps/blob/master/docs/requirements.txt) file for the documentation.

Is there a reason that these dependencies are not listed here:
https://github.com/sandialabs/pvOps/blob/fe6e6579239ab161908fc5b0d3819b720c161f3c/setup.py#L54-L56

I see no reason why they should not be included in the ``DOCS_REQUIRE``, which lets you install documentation dependencies the standard way:

    pip install pvops[docs]

Also, it's HIGHLY recommended to pin the versions of the documentation dependencies as this often ends up causing trouble down the road!

This is part of a JOSS review: https://github.com/openjournals/joss-reviews/issues/5755
