### 2023-09-12 sandialabs/pvOps#60 @AdamRJensen
[JOSS review] Test deprecation warnings
When running the tests, I get 13 warnings. Some of these are fine (e.g., divide by zero).

However, there are some deprecation warnings that should be fixed to avoid issues down the road.

- [ ] FutureWarning: Setting an item of incompatible dtype is deprecated and will raise in a future error of pandas. Value '2013-10-20 00:00:00' has dtype incompatible with float64, please explicitly cast to a compatible dtype first.
    om_df.loc[i, meta_cod] = cod_dates[i]
- [ ] FutureWarning: Series.fillna with 'method' is deprecated and will raise in a future version. Use obj.ffill() or obj.bfill() instead.
    prod_df.loc[:, prod_ener].fillna(method="ffill", inplace=True)
- [ ] MatplotlibDeprecationWarning: The get_cmap function was deprecated in Matplotlib 3.7 and will be removed two minor releases later. Use ``matplotlib.colormaps[name]`` or ``matplotlib.colormaps.get_cmap(obj)`` instead.
    cmap = plt.cm.get_cmap(cmap_name, len(asset_set))
- [ ] DeprecationWarning: module 'sre_constants' is deprecated
    import sre_constants

Feel free to tick off the above boxes as these are fixed.

There's also a conversion warning, but perhaps this is ok?
> RuntimeWarning: some failed to converge after 100 iterations warnings.warn(msg, RuntimeWarning)

This issue is part of a JOSS review: https://github.com/openjournals/joss-reviews/issues/5755
