### 2023-09-12 120981f
Update examples.rst
