+++ docs/pages/examples.rst
@@ -1,5 +1,5 @@
-pvOps Examples
-==============
+Tutorials
+=========
 
 Check out the examples below!
 
