### 2023-09-12 #5755 @AdamRJensen
Comments to the paper from Adam R. Jensen
---------------------------------------------------

- [ ] In the first paragraph of the Statement of Need it would be adviseable to also mention the pvanalytics package and how this fits in with the already mentioned packages.
- [ ] Also, referencing https://openpvtools.readthedocs.io for a list of open source PV packages is highly adviseable
- [ ] Would be nice to give an example of what an operations and maintenance record could contain, e.g., inverter malfunction. Just to make it more concrete for users not too familiar with PV performance data.
- [ ] In the Credit author statement, it would be nice if you also added the _Software_ category as appropriate.
