### 2023-09-12 #5755 @AdamRJensen
Reviewer note
---------------

While I have a lot of insight into the PV world, in particular PV modeling and Python, I'm not exactly the core audience for this package as I typically do not inspect PV performance data.

I'm writing this as I feel like it may be worthwhile to have a person in the core audience review the tutorials? Perhaps @FlorianK13 you represent this group?

I hope it's not inappropriate to bring this up, but please let me know if this is the case. I'm merely interested in delivering the best possible review and ensuring great FOSS in the solar energy domain ❤️ 🌤️
