### 2023-09-15 #5755 @kbonney
Thanks for identifying this need @AdamRJensen. We would certainly appreciate having someone with experience in PV performance and O&M data take a look.

Also, I just want to mention that we are planning on addressing the issues that have been building up on the repository. However it may take a week or two as our team is currently busy with some end of financial year deliverables.
