### 2023-09-15 #5755 @williamhobbs
@AdamRJensen, I am interested in reviewing the tutorials. I've only briefly looked at the readme and docs, so some pointers on where to start and what to test out might be helpful. Also, would I just try things out and provide comments here, or is there a more organized way to go about it?

It is feasible that I could use some real maintenance logs, like in https://github.com/sandialabs/pvOps/blob/master/examples/tutorial_textmodule.ipynb, but I'd have to check on that.

As a very general note about the tutorials, I noticed that there are too many characters on some lines for me to view everything in my browser, e.g.:
