### 2023-09-15 #5755 @AdamRJensen
> @AdamRJensen, I am interested in reviewing the tutorials. I've only briefly looked at the readme and docs, so some pointers on where to start and what to test out might be helpful. Also, would I just try things out and provide comments here, or is there a more organized way to go about it?

I would start with the first listed tutorial and go one by one (that's how I imagine a potential user would interact with it). My feeling was that some of the tutorials lacked sufficient explanation of the individual steps and description of the overall purpose, but that could simply due to my lack of practical experience. Any feedback that you might have, small or large, would be highly appreciated and should go into individual issues in the [pvOps GitHub repository](https://github.com/sandialabs/pvOps/issues/). You can start each issue title with [JOSS review assistance].

> It is feasible that I could use some real maintenance logs, like in https://github.com/sandialabs/pvOps/blob/master/examples/tutorial_textmodule.ipynb, but I'd have to check on that.

If possible, that would be ideal! But if not then that's also ok.
