### 2023-09-21 #5755 @williamhobbs
I just submitted a few issues. I'll add more if I get time to run through more tutorials in detail, but hopefully some of my feedback is general enough to be applied too all of them as appropriate.
