### 2023-10-03 998155d
adding additional docs requirements to setup.py variable
