### 2023-10-03 bb20a42
adding python version in req.txt note
