### 2023-10-03 7cbdd36
adding doc pins to setup py
