### 2023-10-03 0c78f92
updaing rtd yaml to use new doc reqirements location
