### 2023-10-03 89296d0
adding version, copyright, and license to __init__.py
