### 2023-10-03 1df878a
add version to setup.py
