+++ pvops/__init__.py
@@ -3,7 +3,7 @@ from pvops import text2time
 from pvops import timeseries
 from pvops import iv
 
-__version__ = '0.1.0'
+__version__ = '0.2.0'
 
 __copyright__ = """Copyright 2023 National Technology & Engineering 
 Solutions of Sandia, LLC (NTESS). Under the terms of Contract DE-NA0003525 
