### 2023-10-03 b707020
Adding statement of need and accompanying bib items
