### 2023-10-05 040f5e1
updating author, copyright, and version in conf.py and setup.py
