+++ docs/pages/moduleguides/iv.rst
@@ -1,4 +1,4 @@
-IV package
+IV Guide
 ===============
 
 Module Overview
