+++ docs/pages/moduleguides/iv.rst
@@ -8,10 +8,10 @@ Module Overview
 These functions focus on current-voltage (IV) curve simulation and 
 classification.
 
-Examples of usage can be found at:
-  - `tutorial_iv_classifier.ipynb <https://github.com/sandialabs/pvOps/blob/master/examples/tutorial_iv_classifier.ipynb>`_.
-  - `tutorial_iv_diode_extractor.ipynb <https://github.com/sandialabs/pvOps/blob/master/examples/tutorial_iv_diode_extractor.ipynb>`_.
-  - `tutorial_iv_simulator.ipynb <https://github.com/sandialabs/pvOps/blob/master/examples/tutorial_iv_simulator.ipynb>`_.
+Tutorials that exemplify usage can be found at:
+  - `tutorial_iv_classifier.ipynb <https://github.com/sandialabs/pvOps/blob/master/tutorials/tutorial_iv_classifier.ipynb>`_.
+  - `tutorial_iv_diode_extractor.ipynb <https://github.com/sandialabs/pvOps/blob/master/tutorials/tutorial_iv_diode_extractor.ipynb>`_.
+  - `tutorial_iv_simulator.ipynb <https://github.com/sandialabs/pvOps/blob/master/tutorials/tutorial_iv_simulator.ipynb>`_.
 
 extractor
 ^^^^^^^^^^^^^^^^^^^^^
