### 2023-10-05 08911d7
Add Python 3.11 to test environment (#57)
