### 2023-10-05 e945e3b
adding 3.11 to supported versions in docs
