+++ docs/pages/installation.rst
@@ -1,7 +1,7 @@
 Installation
 =============
 
-pvops is tested on Python versions 3.8, 3.9, and 3.10 and depends on a variety of
+pvops is tested on Python versions 3.8, 3.9, 3.10, and 3.11 and depends on a variety of
 packages. See :ref:`reqs` for more information.
 
 The latest release of pvops is accessible via PYPI using the following
