### 2023-10-05 01be577
Addressing various JOSS review comments
* removing unneccesary min requirements file

* Updating RTD docs requirements pins and updated python version used in RTD

* adding additional docs requirements to setup.py variable

* adding python version in req.txt note

* removing docs req txt file

* adding doc pins to setup py

* updaing rtd yaml to use new doc reqirements location

* adding pins to req.txt

* adding version, copyright, and license to __init__.py

* fix citation file

* add version to setup.py

* remove unnecessary init file

* update version to 0.2.0

* removing vestigial files


[…]
