+++ docs/pages/examples.rst
@@ -1,17 +1,17 @@
-pvOps Examples
-==============
+pvOps Tutorials
+===============
 
-Check out the examples below!
+Check out the tutorials below!
 
 .. toctree::
    :maxdepth: 1
-   :caption: List of examples:
+   :caption: List of tutorials:
 
-   examples/tutorial_text2time_module.nblink
-   examples/tutorial_textmodule.nblink
-   examples/tutorial_timeseries.nblink
-   examples/tutorial_AIT_timeseries.nblink
-   examples/tutorial_timeseries_sim.nblink
-   examples/tutorial_iv_simulator.nblink
-   examples/tutorial_iv_classifier.nblink
-   examples/tutorial_iv_diode_extractor.nblink
+   tutorials/tutorial_text2time_module.nblink
+   tutorials/tutorial_textmodule.nblink
+   tutorials/tutorial_timeseries.nblink
+   tutorials/tutorial_AIT_timeseries.nblink
+   tutorials/tutorial_timeseries_sim.nblink
+   tutorials/tutorial_iv_simulator.nblink
+   tutorials/tutorial_iv_classifier.nblink
+   tutorials/tutorial_iv_diode_extractor.nblink
