### 2023-10-06 #5755 @arfon
The JOSS document pipeline is limited in what it can support sorry. While I don't like it, I would suggest either:

1) embedding an image of the table.
2) Trying using a LaTeX table.
3) Restructuring your tables to be less complex.
