### 2023-10-16 bf91ab6
[JOSS review] Update tested versions in documentation (#58)
* Update tested versions in documentation

* adding 3.11 to supported versions in docs

---------

Co-authored-by: Kirk Bonney <47759761+kbonney@users.noreply.github.com>
