### 2023-10-17 1de96db
Adding structure to user guide toctree
