+++ docs/pages/tutorials.rst
@@ -3,15 +3,22 @@ pvOps Tutorials
 
 Check out the tutorials below!
 
-.. toctree::
-   :maxdepth: 1
-   :caption: List of tutorials:
+.. nbgallery::
+   :caption: Text & Text2Time tutorials:
 
-   tutorials/tutorial_text2time_module.nblink
-   tutorials/tutorial_textmodule.nblink
-   tutorials/tutorial_timeseries.nblink
-   tutorials/tutorial_AIT_timeseries.nblink
-   tutorials/tutorial_timeseries_sim.nblink
-   tutorials/tutorial_iv_simulator.nblink
-   tutorials/tutorial_iv_classifier.nblink
-   tutorials/tutorial_iv_diode_extractor.nblink
+   tutorials/tutorial_text2time_module
+   tutorials/tutorial_textmodule
+
+.. nbgallery::
+   :caption: Energy modelling tutorials:
+
+   tutorials/tutorial_timeseries
+   tutorials/tutorial_AIT_timeseries
+   tutorials/tutorial_timeseries_sim
+
+.. nbgallery::
+   :caption: IV tutorials:
+
+   tutorials/tutorial_iv_simulator
+   tutorials/tutorial_iv_classifier
+   tutorials/tutorial_iv_diode_extractor
