### 2023-10-17 fcb6301
Docs updates (#81)
* Adding statement of need and accompanying bib items

* moving statement of need to landing page

* renaming documentation file examples->tutorials

* updating readme to reflect examples->tutorials

* fixing image embedd

* remove reference to  statement of need page

* Adding patch notes
