+++ noxfile.py
@@ -12,7 +12,7 @@ def lint(session):
     """Lint."""
     session.install("flake8")
     session.run("flake8", "--import-order-style", "google")
-    
+
 @nox.session
 def docs(session):
     """Generate documentation."""
@@ -28,13 +28,13 @@ def serve(session):
     except IndexError:
         port = "8085"
     session.run("python", "-m", "http.server", "-b", "localhost", "-d", "docs/_build/html", port)
-   
+
 @nox.session
 def check_style(session):
     """Check if code follows black style."""
     session.install("black")
     session.run("black", "--check", "src")
-    
+
 @nox.session
 def enforce_style(session):
     """Apply black style to code base."""
