### 2023-10-19 506b0eb
Docs updates (#82)
* Adding statement of need and accompanying bib items

* moving statement of need to landing page

* renaming documentation file examples->tutorials

* updating readme to reflect examples->tutorials

* fixing image embedd

* remove reference to  statement of need page

* Adding patch notes

* removing WIP docs pages

* Remvoing reference to undefined reference "reqs"

* Fix image embed warnings

* adding noxfile for taskrunning

* Expanding temrinology page

* Adding structure to user guide toctree

* restructuring index

* Adding gallery to tutorials page

* updating doc pi
[…]
