+++ docs/pages/abbreviations.rst
@@ -1,20 +1,17 @@
-Abbreviations
+Abbreviations/Terminology
 ====================================
-* AIT:
+* AIT: Additive Interaction Model described in :cite:p:`app12041872`
 * CEC: California Energy Commission
 * WS: wind speed
-* Varr
-* prod_irradiance_filter
-* popular
-* T
-* Rsh_mult
-* Rs_mult
-* Io_mult
-* Il_mult
-* nnsvth_mult
-* E
-* Tc
-* mode
-* gt
-* v_rbd
-* v_oc
+* Varr: Voltage array
+* T: Average cell temperature
+* Rsh_mult: Multiplier usually less than 1 to simulate a drop in RSH
+* Rs_mult: Multiplier usually less than 1 to simulate a drop in RS
+* Io_mult: Multiplier usually less than 1 to simulate a drop in IO
+* Il_mult: Multiplier usually less than 1 to simulate a drop in IL
+* nnsvth_mult: Multiplier usually less than 1 to simulate a drop in NNSVTH
+* E: Irradiance 
+* Tc: Cell temp
+* gt: (G - Irradiation and T - temperature)
+* v_rbd: Reverse bias diode voltage 
+* v_oc: Open circuit voltage
