+++ .github/workflows/builddocs.yml
@@ -28,8 +28,8 @@ jobs:
         run: sudo apt-get update -y && sudo apt-get install -y pandoc
       - uses: actions/setup-python@v2
         with: 
-          python-version: '3.8' 
-      - name: Setup Python 3.8
+          python-version: '3.11' 
+      - name: Setup Python 3.11
         run: |
           python -m pip install --upgrade pip
           python -m pip install -r requirements.txt
