### 2023-10-24 5b19da3
attempting latex table
