### 2023-10-24 #5755 @kbonney
> The JOSS document pipeline is limited in what it can support sorry. While I don't like it, I would suggest either:
>
> 1. embedding an image of the table.
> 2. Trying using a LaTeX table.
> 3. Restructuring your tables to be less complex.

Hi @arfon. I remade the table in LaTeX, but I am unsure how to correctly pull it into the JOSS document. I just tried copying and pasting the raw latex into the file, but that did not seem to work. See [here](https://github.com/kbonney/pvOps/actions/runs/6632354386).  There are some dependencies that need to be imported so I am sure that is part of the issue, but I am not sure how to do that with the inara workflow.
