### 2023-10-24 536ebdf
reverting to simplified markdown style table
