### 2023-10-24 86ad883
Tutorial and Example bug fixes (#83)
* adressing bug in tutorial

* addressing future warning

* addressing test warnings
