+++ docs/pages/moduleguides/timeseries.rst
@@ -13,7 +13,7 @@ types of expected energy models:
 * Linear: a high flexibility linear regression model.
 
 Additionally, the ability to generate expected energy via IEC 
-standards is implemented in the :py:mod:`~pvops.timeseries.models.iec`
+standards (iec 61724-1) is implemented in the :py:mod:`~pvops.timeseries.models.iec`
 module.
 
 An example of usage can be found in 
