### 2023-10-24 1ad7d73
fixing joss pub table (#84)
* Updating contributing page and adding a few lines to gitignore

* Small edit

* adding files to paper folder for JOSS

* Adding joss workflow

* Making adjustments to get joss workflow to run

* Minor updates to bib formatting and spacing in paper

* Attempt at fixing table newlines

* Another attempt

* Attempt 3

* Now we are on track

* Trying html unordered lists

* testing a new table format

* try again

* attempting a picture approach

* Adding JOSS status

* Updating python setup action in testing workflow

* Keyword search in text from reference dictionar
[…]
