+++ pvops/__init__.py
@@ -2,3 +2,11 @@ from pvops import text
 from pvops import text2time
 from pvops import timeseries
 from pvops import iv
+
+__version__ = '0.2.0'
+
+__copyright__ = """Copyright 2023 National Technology & Engineering 
+Solutions of Sandia, LLC (NTESS). Under the terms of Contract DE-NA0003525 
+with NTESS, the U.S. Government retains certain rights in this software."""
+
+__license__ = "BSD 3-Clause License"
