+++ docs/pages/classguides/modcell.rst
@@ -1,3 +1,0 @@
-
-Modcell Guide
-=============
