+++ docs/requirements.txt
@@ -1,9 +1,0 @@
-ipykernel
-nbsphinx
-nbsphinx-link
-sphinx-copybutton
-sphinx_rtd_theme
-sphinxcontrib.bibtex
-nbformat
-nbconvert
-coverage

[…]
