### 2023-10-24 aa4ecd2
Tutorials updates (#85)
* adding link to tutorials in README

* specifying plane of array irradiance

* specifying plane of array irradiance

* added iec standard number

* remove redundant plots and add data description
