### 2023-11-03 6740402
cleaning up bipartite graphing
