### 2023-11-03 44de392
cleaning up vis function readability
