### 2023-11-03 cbaebb6
reverting to returning fig rather than ax to be consistent with the rest of the repo
