### 2023-11-04 #5755 @AdamRJensen
> # Review comments from Adam R. Jensen
>
> Comments to the paper
> --------------------------
>
> - [ ] In the first paragraph of the Statement of Need it would be adviseable to also mention the pvanalytics package and how this fits in with the already mentioned packages.
> - [ ] Also, referencing https://openpvtools.readthedocs.io for a list of open source PV packages is highly advisable
> - [ ] It would be nice to give an example of what an operations and maintenance record could contain, e.g., inverter malfunction. Just to make it more concrete for users not too familiar with PV performance data.
> - [ ] In the Credit author statement, it would be nice if you also added the _Software_ category as appropriate.
> - [ ] I'm missing a concrete example in the Statement of Need. Here's a hypothetical example "Failures in PV power plants are noted in O&M logs, which often are overlooked due to difficulty in parsing these files, however it's important to understand what causes plant failures, e.g., inverter breakdown, jumping goats. pvops contains functions to solve this bla bla bla"
> - [ ] Add a caption to Table 1.
> - [ ] Include one image from one of the tutorials in the paper as a visual example (optional)

@kbonney did you address these comments to the paper?
