### 2023-11-06 39f933c
Bipartite graphing for attribute connectivity (#86)
Updating visualize_attribute_connectivity to use bipartite graph layout

---------

Co-authored-by: Benjamin Freixas Emery <bfemery@sandia.gov>
