+++ pvops/text/visualize.py
@@ -3,6 +3,7 @@ import matplotlib
 import matplotlib.pyplot as plt
 import networkx as nx
 from sklearn.metrics import ConfusionMatrixDisplay
+from networkx.algorithms import bipartite
 
 # data structures
 import numpy as np
@@ -23,13 +24,16 @@ from nltk.tokenize import word_tokenize
 def visualize_attribute_connectivity(
     om_df,
     om_col_dict,
-    figsize=(40, 20),
+    figsize=(20, 10),
     attribute_colors=["lightgreen", "cornflowerblue"],
     edge_width_scalar=10,
     graph_aargs={},
 ):
     """Visualize a knowledge graph which shows the frequency of combinations between attributes
     ``ATTRIBUTE1_COL`` and ``ATTRIBUTE2_COL``
+
+    NOW USES BIPARTITE LAYOUT
+    ATTRIBUTE2_COL is colored using a colormap.
 
     Parameters
     ----------
@@ -45,12 +49,12 @@ def visualize_attribute_connectivity(
             }
 
     figsize : tuple
-        Figure size
-    attribute_colors : list
+        Figure size, defaults to (20,10)
+    attribute_colors : list[str]
         List of two strings which designate the colors for Attribute1 and Attribute 2, respectively.
     edge_width_scalar : numeric
-        Weight utilized to cause dynamic widths based on number of connections between Attribute 1
-        and Attribute 2.
+        Weight utilized to scale widths based on number of connections between Attribute 1
+        and Attribute 2. Larger values will produce larger widths, and smaller values will produce smaller widths.
     graph_aargs : dict
         Optional, arguments passed to networkx graph drawer.
         Suggested attributes to pass:
@@ -59,38 +63,57 @@ def visualize_attribute_connectivity(
         - font_weight='bold'
         - node_size=19000
         - font_size=35
-        - node_color='darkred'
-        - font_color='red'
 
     Returns
     -------
-    Matplotlib figure instance,
-    networkx EdgeView object
-    i.e. [('A', 'X'), ('X', 'B'), ('C', 'Y'), ('C', 'Z')]
+    Matplotlib axis,
+    networkx graph
     """
-    df = om_df.copy()
+    # initialize figure
+    fig = plt.figure(figsize=figsize, facecolor='w', edgecolor='k')
+    ax = plt.gca()
+
+    # attribute column names
     ATTRIBUTE1_COL = om_col_dict["attribute1_col"]
     ATTRIBUTE2_COL = om_col_dict["attribute2_col"]
 
-    nx_data = []
-    for a in np.unique(df[ATTRIBUTE1_COL].tolist()):
-        df_iter = df[df[ATTRIBUTE1_COL] == a]
-        for i in np.unique(df_iter[ATTRIBUTE2_COL].tolist()):
-            w = len(df_iter[df_iter[ATTRIBUTE2_COL] == i])
-            nx_data.append([a, i, w])
-
-    unique_df = pd.DataFrame(
-        nx_data, columns=[ATTRIBUTE1_COL, ATTRIBUTE2_COL, "w"])
-
-    G = nx.from_pandas_edgelist(unique_df, ATTRIBUTE1_COL, ATTRIBUTE2_COL, "w")
-    fig = plt.figure(figsize=figsize)
-    fig.suptitle(
-        f"Connectivity between {ATTRIBUTE1_COL} and {ATTRIBUTE2_COL}",
-        fontsize=50,
-        y=1.08,
+    ax.set_title(
+        f"Connectivity between {ATTRIBUTE2_COL} and {ATTRIBUTE1_COL}",
         fontweight="bold",
     )
 
[…]
