### 2023-11-06 0edee50
joss paper updates based on adam's comments
