### 2023-11-06 534cf88
Joss paper updates (#87)
Paper updates:

Added software to credit author statement
Added example of maintenance record use
mention openpvtools
mention pvanalytics
caption addressed
