+++ paper.bib
@@ -185,3 +185,10 @@
   volume={12},
   pages={2825--2830}
 }
+
+@techreport{perry2022pvanalytics,
+  title={PVAnalytics: A Python Package for Automated Processing of Solar Time Series Data},
+  author={Perry, Kirsten and Vining, William and Anderson, Kevin and Muller, Matthew and Hansen, Cliff},
+  year={2022},
+  institution={National Renewable Energy Lab.(NREL), Golden, CO (United States)}
+}
