### 2023-11-07 b27b21e
minor edits to text
