### 2023-11-07 fef96ad
minor edits to text (#89)
* minor edits to text

* removing redundant reference to openpvtools

---------

Co-authored-by: Kirk Bonney <47759761+kbonney@users.noreply.github.com>
