### 2023-11-07 c6924de
Update paper.md
fix paper caption
