### 2023-11-08 4c7c6df
Update paper.md
fixing some punctuation
