### 2023-11-08 #5755 @rkurchin
Some small editorial comments:
* line 29: remove comma after "open-source"
* line 32: are timeseries and text records the only data types? If not, "i.e." should be "e.g." (similarly on the next line for completeness and consistency as the only indicators of quality)
* Regarding formatting for table 1, perhaps add lines between each row to more clearly separate the entries in the last column? Also, there's an errant period in the bottom row of the last column
* I'd suggest breaking out the numbered list on lines 53-62 onto separate lines to improve readability
* line 64: remove comma after "capabilities"
