+++ .github/workflows/lint-and-test.yml
@@ -18,7 +18,7 @@ jobs:
     - name: Install pvops
       run: |
         python -m pip install --upgrade pip
-        pip install .
+        pip install .[iv]
     - name: Test with pytest
       run: |
         pip install pytest pytest-cov
