### 2023-11-09 a888c40
Make IV capabilities optional (#90)
* making iv optional

* add iv option to testing workflow
