+++ .readthedocs.yml
@@ -15,4 +15,4 @@ python:
     - method: pip
       path: .
       extra_requirements:
-        - doc
+        - doc, iv
