### 2023-11-09 158f300
Update whatsnew (#91)
* updating whatsnew

* updating version
