+++ paper.bib
@@ -167,11 +167,13 @@
   publisher = {O'Reilly Media}
 }
 
-@misc{chollet2015keras,
-  title={Keras},
-  author={Chollet, Fran\c{c}ois and others},
-  year={2015},
-  howpublished={\url{https://keras.io}}
+@misc{chollet2015,
+author = {François Chollet },
+title = {keras},
+year = {2015},
+publisher = {GitHub},
+journal = {GitHub repository},
+howpublished = {\url{https://github.com/fchollet/keras}}
 }
 
 @article{pedregosa2011sklearn,
