### 2023-11-15 #5755 @kbonney
@rkurchin
I searched around and it doesn't look like keras has an official DOI. There is a conversation [here](https://github.com/keras-team/keras/issues/247) about it in their issues. If having a DOI is a strict requirement to move forward, we could switch the citation to the backend library that keras uses called tensorflow or just drop the citation all together. If the package doesn't have a proper DOI, I take that as it not being important to the maintainer to be cited...
