### 2023-11-15 #5755 @kyleniemeyer
We only ask people to include DOIs for things that have DOIs—if the software doesn't have one, then we certainly can't require the authors to cite one somehow. (Although I have seen JOSS authors poke maintainers to make it easier to cite their software, with some success, since it's a pretty light lift.)

Instead, just cite the software you reference in the article to the best of your ability. Something is better than nothing.
