### 2023-11-15 #5755 @rkurchin
Makes sense, then I think this is good to go – but you might have to do some magic overwrite or something on the acceptance checks if that DOI thing is triggered...
