### 2023-11-16 8a04559
Adds missing } in citation
