+++ paper.bib
@@ -172,6 +172,7 @@
   author={Chollet, Fran\c{c}ois and others},
   year={2015},
   howpublished={\url{https://keras.io}}
+}
 
 @article{pedregosa2011sklearn,
   title={Scikit-learn: Machine Learning in {P}ython},
