### 2023-11-16 #5755 @kyleniemeyer
Hi @kbonney, can you merge this PR? The keras citation was missing a } to end the entry: https://github.com/sandialabs/pvOps/pull/92
