### 2023-11-16 #5755 @kbonney
> Hi @kbonney, can you merge this PR? The keras citation was missing a } to end the entry: [sandialabs/pvOps#92](https://github.com/sandialabs/pvOps/pull/92)

Merged. Thanks for catching that.
