### 2023-11-16 81efb60
Light edits in JOSS paper (#93)
Mostly adds missing commas
