### 2023-11-16 #5755 @kyleniemeyer
Congratulations @kbonney on your article's publication in JOSS! Please consider [signing up to review for JOSS](https://reviewers.joss.theoj.org/join) if you haven't already.

Many thanks to @FlorianK13 and @AdamRJensen for reviewing this submission, and to @rkurchin for editing—we can't do it without you!
